+++ app/Main.hs
@@ -3,6 +3,7 @@ module Main where
 import Control.Varying
 import Control.Applicative
 import Text.Printf
+import Data.Functor.Identity
 
 -- | A simple 2d point type.
 data Point = Point { px :: Float
@@ -13,7 +14,7 @@ data Point = Point { px :: Float
 -- loops forever. This spline takes float values of delta time as input,
 -- outputs the current x value at every step and would result in () if it
 -- terminated.
-tweenx :: (Applicative m, Monad m) => Spline Float Float m ()
+tweenx :: (Applicative m, Monad m) => SplineT Float Float m ()
 tweenx = do
     -- Tween from 0 to 100 over 1 second
     x <- tween easeOutExpo 0 100 1
@@ -24,24 +25,24 @@ tweenx = do
 
 -- A quadratic tween back and forth from 0 to 100 over 2 seconds that never
 -- ends.
-tweeny :: (Applicative m, Monad m) => Spline Float Float m ()
+tweeny :: (Applicative m, Monad m) => SplineT Float Float m ()
 tweeny = do
     y <- tween easeOutQuad 0 100 1
     _ <- tween easeOutQuad y 0 1
     tweeny
 
 -- Our time signal that provides delta time samples.
-time :: Var IO a Float
+time :: VarT IO a Float
 time = deltaUTC
 
 -- | Our Point value that varies over time continuously in x and y.
-backAndForth :: Var IO a Point
+backAndForth :: VarT IO a Point
 backAndForth =
-    -- Turn our splines back into continuous value streams. We must provide
+    -- Turn our splines into continuous output streams. We must provide
     -- a starting value since splines are not guaranteed to be defined at
     -- their edges.
-    let x = execSpline 0 tweenx
-        y = execSpline 0 tweeny
+    let x = outputStream 0 tweenx
+        y = outputStream 0 tweeny
     in
     -- Construct a varying Point that takes time as an input.
     (Point <$> x <*> y)
@@ -53,10 +54,13 @@ backAndForth =
 
 main :: IO ()
 main = do
-    putStrLn "Varying Values"
+    putStrLn "An example of value streams using the varying library."
+    putStrLn "Enter a newline to continue, quit with ctrl+c"
+    _ <- getLine
+
     loop backAndForth
-        where loop :: Var IO () Point -> IO ()
-              loop v = do (point, vNext) <- runVar v ()
+        where loop :: VarT IO () Point -> IO ()
+              loop v = do (point, vNext) <- runVarT v ()
                           printf "\nPoint %03.1f %03.1f" (px point) (py point)
                           loop vNext