+++ src/Control/Varying/Core.hs
@@ -7,12 +7,13 @@
 --   Value streams represent values that change over a given domain.
 --
 --   A stream takes some input (the domain e.g. time, place, etc) and when
---   sampled using 'runVar' - produces a value and a new value stream. This
+--   sampled using 'runVarT' - produces a value and a new value stream. This
 --   pattern is known as an automaton. `varying` uses this pattern as its base
 --   type with the additon of a monadic computation to create locally stateful
 --   signals that change over some domain.
 module Control.Varying.Core (
-    Var(..),
+    Var,
+    VarT(..),
     -- * Creating value streams
     -- $creation
     var,
@@ -33,6 +34,7 @@ module Control.Varying.Core (
     loopVar_,
     whileVar,
     whileVar_,
+    scanVar,
     stepMany,
     -- * Testing value streams
     testVar,
@@ -46,9 +48,10 @@ module Control.Varying.Core (
 import Prelude hiding (id, (.))
 import Control.Arrow
 import Control.Category
-import Control.Monad (when)
+import Control.Monad 
 import Control.Applicative
 import Data.Monoid
+import Data.Functor.Identity
 import Debug.Trace
 --------------------------------------------------------------------------------
 -- $creation
@@ -56,7 +59,7 @@ import Debug.Trace
 -- with 'var':
 --
 -- @
--- addsOne :: Monad m => Var m Int Int
+-- addsOne :: Monad m => VarT m Int Int
 -- addsOne = var (+1)
 -- @
 --
@@ -66,7 +69,7 @@ import Debug.Trace
 -- @(a -> m b)@ using 'varM':
 --
 -- @
--- getsFile :: Var IO FilePath String
+-- getsFile :: VarT IO FilePath String
 -- getsFile = varM readFile
 -- @
 --
@@ -75,127 +78,134 @@ import Debug.Trace
 -- over how value streams are stepped and sampled:
 --
 -- @
--- delay :: Monad m => b -> Var m a b -> Var m a b
--- delay b v = Var $ \a -> return (b, go a v)
---     where go a v' = Var $ \a' -> do (b', v'') <- runVar v' a
+-- delay :: Monad m => b -> VarT m a b -> VarT m a b
+-- delay b v = VarT $ \a -> return (b, go a v)
+--     where go a v' = VarT $ \a' -> do (b', v'') <- runVarT v' a
 --                                     return (b', go a' v'')
 -- @
 --
 --------------------------------------------------------------------------------
--- | Lift a pure computation into a 'Var'.
-var :: Applicative a => (b -> c) -> Var a b c
-var f = Var $ \a -> pure (f a, var f)
+-- | Lift a pure computation into a stream.
+var :: Applicative m => (a -> b) -> VarT m a b
+var f = VarT $ \a -> pure (f a, var f)
 
--- | Lift a monadic computation into a 'Var'.
-varM :: Monad m => (a -> m b) -> Var m a b
-varM f = Var $ \a -> do
+-- | Lift a monadic computation into a stream.
+varM :: Monad m => (a -> m b) -> VarT m a b
+varM f = VarT $ \a -> do
     b <- f a
     return (b, varM f)
 
--- | Create a 'Var' from a state transformer.
+-- | Create a stream from a state transformer.
 mkState :: Monad m
         => (a -> s -> (b, s)) -- ^ state transformer
         -> s -- ^ intial state
-        -> Var m a b
-mkState f s = Var $ \a -> do
+        -> VarT m a b
+mkState f s = VarT $ \a -> do
   let (b', s') = f a s
   return (b', mkState f s')
 --------------------------------------------------------------------------------
 -- $running
 -- The easiest way to sample a stream is to run it in the desired monad with
--- 'runVar'. This will produce a sample value and a new stream.
+-- 'runVarT'. This will produce a sample value and a new stream.
 --
--- > do (sample, v') <- runVar v inputValue
+-- > do (sample, v') <- runVarT v inputValue
 --
 -- Much like Control.Monad.State there are other entry points for running
 -- value streams like 'evalVar', 'execVar'. There are also extra control
 -- structures such as 'loopVar' and 'whileVar'.
 --------------------------------------------------------------------------------
--- | Iterate a 'Var' once and return the sample value.
-evalVar :: Functor m => Var m a b -> a -> m b
-evalVar v a = fst <$> runVar v a
+-- | Iterate a stream once and return the sample value.
+evalVar :: Functor m => VarT m a b -> a -> m b
+evalVar v a = fst <$> runVarT v a
 
--- | Iterate a 'Var' once and return the next 'Var'.
-execVar :: Functor m => Var m a b -> a -> m (Var m a b)
-execVar v a = snd <$> runVar v a
+-- | Iterate a stream once and return the next stream.
+execVar :: Functor m => VarT m a b -> a -> m (VarT m a b)
+execVar v a = snd <$> runVarT v a
 
--- | Loop over a 'Var' that takes no input value.
-loopVar_ :: (Functor m, Monad m) => Var m () a -> m ()
+-- | Loop over a stream that takes no input value.
+loopVar_ :: (Functor m, Monad m) => VarT m () a -> m ()
 loopVar_ v = execVar v () >>= loopVar_
 
--- | Loop over a 'Var' that produces its own next input value.
-loopVar :: Monad m => a -> Var m a a -> m a
-loopVar a v = runVar v a >>= uncurry loopVar
+-- | Loop over a stream that produces its own next input value.
+loopVar :: Monad m => a -> VarT m a a -> m a
+loopVar a v = runVarT v a >>= uncurry loopVar
 
--- | Iterate a 'Var' that requires no input until the given predicate fails.
-whileVar_ :: Monad m => (a -> Bool) -> Var m () a -> m a
+-- | Iterate a stream that requires no input until the given predicate fails.
+whileVar_ :: Monad m => (a -> Bool) -> VarT m () a -> m a
 whileVar_ f v = do
-   (a, v') <- runVar v ()
+   (a, v') <- runVarT v ()
    if f a then whileVar_ f v' else return a
 
--- | Iterate a 'Var' that produces its own next input value until the given
+-- | Iterate a stream that produces its own next input value until the given
 -- predicate fails.
 whileVar :: Monad m
          => (a -> Bool) -- ^ The predicate to evaluate samples.
          -> a -- ^ The initial input/sample value.
-         -> Var m a a -- ^ The 'Var' to iterate
+         -> VarT m a a -- ^ The stream to iterate
          -> m a -- ^ The last sample
 whileVar f a v = if f a
-                 then runVar v a >>= uncurry (whileVar f)
+                 then runVarT v a >>= uncurry (whileVar f)
                  else return a
 
--- | Iterate a 'Var' using a list of input until all input is consumed and
+-- | Iterate a stream using a list of input until all input is consumed and
 -- output the result.
-stepMany :: (Monad m, Functor m, Monoid a) => [a] -> Var m a b -> m (b, Var m a b)
-stepMany ([e]) y = runVar y e
+stepMany :: (Monad m, Functor m, Monoid a) => [a] -> VarT m a b -> m (b, VarT m a b)
+stepMany ([e]) y = runVarT y e
 stepMany (e:es) y = execVar y e >>= stepMany es
-stepMany []     y = runVar y mempty
+stepMany []     y = runVarT y mempty
+
+-- | Run the stream over the input values, gathering the output values in a 
+-- list. 
+scanVar :: (Applicative m, Monad m) => VarT m a b -> [a] -> m [b]
+scanVar v = liftM snd . foldM f (v,[])
+    where f (v', outs) a = do (b, v'') <- runVarT v' a
+                              return (v'', outs ++ [b])
 --------------------------------------------------------------------------------
 -- Testing and debugging
 --------------------------------------------------------------------------------
--- | Trace the sample value of a 'Var' and pass it along as output. This is
--- very useful for debugging graphs of 'Var's.
-vtrace :: (Applicative a, Show b) => Var a b b
+-- | Trace the sample value of a stream and pass it along as output. This is
+-- very useful for debugging graphs of streams.
+vtrace :: (Applicative a, Show b) => VarT a b b
 vtrace = vstrace ""
 
--- | Trace the sample value of a 'Var' with a prefix and pass the sample along
--- as output. This is very useful for debugging graphs of 'Var's.
-vstrace :: (Applicative a, Show b) => String -> Var a b b
+-- | Trace the sample value of a stream with a prefix and pass the sample along
+-- as output. This is very useful for debugging graphs of streams.
+vstrace :: (Applicative a, Show b) => String -> VarT a b b
 vstrace s = vftrace ((s ++) . show)
 
 -- | Trace the sample value after being run through a "show" function.
--- This is very useful for debugging graphs of 'Var's.
-vftrace :: Applicative a => (b -> String) -> Var a b b
+-- This is very useful for debugging graphs of streams.
+vftrace :: Applicative a => (b -> String) -> VarT a b b
 vftrace f = var $ \b -> trace (f b) b
 
--- | A utility function for testing 'Var's that don't require input. Runs
--- a 'Var' printing each sample until the given predicate fails.
-testWhile_ :: Show a => (a -> Bool) -> Var IO () a -> IO ()
+-- | A utility function for testing streams that don't require input. Runs
+-- a stream printing each sample until the given predicate fails.
+testWhile_ :: Show a => (a -> Bool) -> VarT IO () a -> IO ()
 testWhile_ f v = do
-    (a, v') <- runVar v ()
+    (a, v') <- runVarT v ()
     when (f a) $ print a >> testWhile_ f v'
 
--- | A utility function for testing 'Var's that require input. The input
--- must have a 'Read' instance. Use this in GHCI to step through your 'Var's
+-- | A utility function for testing streams that require input. The input
+-- must have a 'Read' instance. Use this in GHCI to step through your streams
 -- by typing the input and hitting `return`.
-testVar :: (Read a, Show b) => Var IO a b -> IO ()
+testVar :: (Read a, Show b) => VarT IO a b -> IO ()
 testVar v = loopVar_ $ varM (const $ putStrLn "input: ")
                     ~> varM (const getLine)
                     ~> var read
                     ~> v
                     ~> varM print
 
--- | A utility function for testing 'Var's that don't require input. Use
--- this in GHCI to step through your 'Var's using the `return` key.
-testVar_ :: Show b => Var IO () b -> IO ()
+-- | A utility function for testing streams that don't require input. Use
+-- this in GHCI to step through your streams using the `return` key.
+testVar_ :: Show b => VarT IO () b -> IO ()
 testVar_ v = loopVar_ $ pure () ~> v ~> varM print ~> varM (const getLine)
 --------------------------------------------------------------------------------
 -- Adjusting and accumulating
 --------------------------------------------------------------------------------
 -- | Accumulates input values using a folding function and yields
 -- that accumulated value each sample.
-accumulate :: Monad m => (c -> b -> c) -> c -> Var m b c
-accumulate f b = Var $ \a -> do
+accumulate :: Monad m => (c -> b -> c) -> c -> VarT m b c
+accumulate f b = VarT $ \a -> do
     let b' = f b a
     return (b', accumulate f b')
 
@@ -204,10 +214,10 @@ accumulate f b = Var $ \a -> do
 -- themselves for values. For example:
 --
 -- > let v = 1 + delay 0 v in testVar_ v
-delay :: Monad m => b -> Var m a b -> Var m a b
-delay b v = Var $ \a -> return (b, go a v)
-    where go a v' = Var $ \a' -> do (b', v'') <- runVar v' a
-                                    return (b', go a' v'')
+delay :: Monad m => b -> VarT m a b -> VarT m a b
+delay b v = VarT $ \a -> return (b, go a v)
+    where go a v' = VarT $ \a' -> do (b', v'') <- runVarT v' a
+                                     return (b', go a' v'')
 --------------------------------------------------------------------------------
 -- $composition
 -- You can compose value streams together using '~>' and '<~'. The "right plug"
@@ -217,27 +227,27 @@ delay b v = Var $ \a -> return (b, go a v)
 -- streams that read naturally.
 --------------------------------------------------------------------------------
 -- | Same as '~>' with flipped parameters.
-(<~) :: Monad m => Var m b c -> Var m a b -> Var m a c
+(<~) :: Monad m => VarT m b c -> VarT m a b -> VarT m a c
 (<~) = flip (~>)
 infixl 1 <~
 
--- | Connects two 'Var's by chaining the first's output into the input of the
--- second. This is the defacto 'Var' composition method and in fact '.' is an
+-- | Connects two streams by chaining the first's output into the input of the
+-- second. This is the defacto stream composition method and in fact '.' is an
 -- alias of '<~', which is just '~>' flipped.
-(~>) :: Monad m => Var m a b -> Var m b c -> Var m a c
-(~>) v1 v2 = Var $ \a -> do
-    (b, v1') <- runVar v1 a
-    (c, v2') <- runVar v2 b
+(~>) :: Monad m => VarT m a b -> VarT m b c -> VarT m a c
+(~>) v1 v2 = VarT $ \a -> do
+    (b, v1') <- runVarT v1 a
+    (c, v2') <- runVarT v2 b
     return (c, v1' ~> v2')
 infixr 1 ~>
 --------------------------------------------------------------------------------
 -- Typeclass instances
 --------------------------------------------------------------------------------
--- | You can transform the sample value of any 'Var':
+-- | You can transform the sample value of any stream:
 --
 -- >  fmap (*3) $ accumulate (+) 0
 -- Will sum input values and then multiply the sum by 3.
-instance (Applicative m, Monad m) => Functor (Var m b) where
+instance (Applicative m, Monad m) => Functor (VarT m b) where
     fmap f' v = v ~> var f'
 
 -- | A very simple category instance.
@@ -252,20 +262,20 @@ instance (Applicative m, Monad m) => Functor (Var m b) where
 --
 -- It is preferable for consistency (and readability) to use 'plug left' ('<~')
 -- and 'plug right' ('~>') instead of ('.') where possible.
-instance (Applicative m, Monad m) => Category (Var m) where
+instance (Applicative m, Monad m) => Category (VarT m) where
     id = var id
     f . g = g ~> f
 
--- | 'Var's are applicative.
+-- | Streams are applicative.
 --
 -- >  (,) <$> pure True <*> var "Applicative"
-instance (Applicative m, Monad m) => Applicative (Var m a) where
+instance (Applicative m, Monad m) => Applicative (VarT m a) where
     pure = var . const
-    vf <*> va = Var $ \a -> do (f, vf') <- runVar vf a
-                               (b, va') <- runVar va a
-                               return (f b, vf' <*> va')
+    vf <*> va = VarT $ \a -> do (f, vf') <- runVarT vf a
+                                (b, va') <- runVarT va a
+                                return (f b, vf' <*> va')
 
--- | 'Var's are arrows, which means you can use proc notation.
+-- | Streams are arrows, which means you can use proc notation.
 --
 -- @
 -- v = proc a -> do
@@ -276,23 +286,23 @@ instance (Applicative m, Monad m) => Applicative (Var m a) where
 -- which is equivalent to
 --
 -- >  v = (\ex ey -> (+) <$> ex <*> ey) <$> intEventVar <*> anotherIntEventVar
-instance (Applicative m, Monad m) => Arrow (Var m) where
+instance (Applicative m, Monad m) => Arrow (VarT m) where
     arr = var
-    first v = Var $ \(b,d) -> do (c, v') <- runVar v b
-                                 return ((c,d), first v')
+    first v = VarT $ \(b,d) -> do (c, v') <- runVarT v b
+                                  return ((c,d), first v')
 
--- | 'Var's can be monoids
+-- | Streams can be monoids
 --
 -- > let v = var (const "Hello ") `mappend` var (const "World!")
-instance (Applicative m, Monad m, Monoid b) => Monoid (Var m a b) where
+instance (Applicative m, Monad m, Monoid b) => Monoid (VarT m a b) where
     mempty = pure mempty
     mappend = liftA2 mappend
 
--- | 'Var's can be written as numbers.
+-- | Streams can be written as numbers.
 --
 -- >  let v = 1 ~> accumulate (+) 0
 -- which will sum the natural numbers.
-instance (Applicative m, Monad m, Num b) => Num (Var m a b) where
+instance (Applicative m, Monad m, Num b) => Num (VarT m a b) where
     (+) = liftA2 (+)
     (-) = liftA2 (-)
     (*) = liftA2 (*)
@@ -300,11 +310,11 @@ instance (Applicative m, Monad m, Num b) => Num (Var m a b) where
     signum = fmap signum
     fromInteger = pure . fromInteger
 
--- | 'Var's can be written as floats.
+-- | Streams can be written as floats.
 --
 -- >  let v = pi ~> accumulate (*) 0.0
 -- which will attempt (and succeed) to multiply pi by zero every step.
-instance (Applicative m, Monad m, Floating b) => Floating (Var m a b) where
+instance (Applicative m, Monad m, Floating b) => Floating (VarT m a b) where
     pi = pure pi
     exp = fmap exp
     log = fmap log
@@ -312,23 +322,28 @@ instance (Applicative m, Monad m, Floating b) => Floating (Var m a b) where
     cos = fmap cos; cosh = fmap cosh; acos = fmap acos; acosh = fmap acosh
     atan = fmap atan; atanh = fmap atanh
 
--- | 'Var's can be written as fractionals.
+-- | Streams can be written as fractionals.
 --
 -- >  let v = 2.5 ~> accumulate (+) 0
 -- which will add 2.5 each step.
-instance (Applicative m, Monad m, Fractional b) => Fractional (Var m a b) where
+instance (Applicative m, Monad m, Fractional b) => Fractional (VarT m a b) where
     (/) = liftA2 (/)
     fromRational = pure . fromRational
 --------------------------------------------------------------------------------
 -- Core datatypes
 --------------------------------------------------------------------------------
--- | The vessel of a value stream. A 'Var' is a structure that contains a value
--- that changes over some input. That input could be time (Float, Double, etc)
--- or 'Control.Varying.Event.Event's or 'Char' - whatever.
--- It's a kind of Mealy machine (an automaton) with effects.
-data Var m b c =
-     Var { runVar :: b -> m (c, Var m b c)
-                  -- ^ Given an input value, return a computation that
-                  -- effectfully produces an output value (a sample) and a 'Var'
-                  -- for producing the next sample.
-         }
+-- | A value stream parameterized with Identity that takes input of type @a@
+-- and gives output of type @b@. This is the pure, effect-free version of
+-- 'VarT'.
+type Var a b = VarT Identity a b
+
+-- | A value stream is a structure that contains a value that changes over some 
+-- input. It's a kind of Mealy machine (an automaton) with effects. Using
+-- 'runVarT' with an input value of type 'a' yields a "step", which is a value 
+-- of type 'b' and a new 'VarT' for yielding the next value.
+data VarT m a b =
+     VarT { runVarT :: a -> m (b, VarT m a b)
+            -- ^ Given an input value, return a computation that effectfully 
+            -- produces an output value and a new stream for producing the next 
+            -- sample.
+          }