+++ changelog.md
@@ -7,4 +7,8 @@ change log
 
 0.3.0.0 - updated the type of mapOutput to a more friendly, usable signature
           bug fixes
-0.3.1.0 - added stepMany
+
+0.3.1.0 - added stepMany, eitherE
+
+0.4.0.0 - Var and Spline are now parameterized with Identity, removed mix, changed
+          the behavior of race, added untilEvent variants, added tests.